+++ exporter/lokiexporter/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/gogo/protobuf v1.3.2
 	github.com/golang/snappy v0.0.4
 	github.com/mattn/go-colorable v0.1.7 // indirect
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil v0.0.0-00010101000000-000000000000
 	github.com/prometheus/common v0.30.0
 	github.com/prometheus/prometheus v1.8.2-0.20210621150501-ff58416a0b02
 	github.com/stretchr/testify v1.7.0
@@ -14,3 +15,5 @@ require (
 	go.uber.org/zap v1.19.0
 	google.golang.org/grpc v1.40.0
 )
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil => ../../internal/testutil
