+++ extension/httpforwarder/go.mod
@@ -3,8 +3,11 @@ module github.com/open-telemetry/opentelemetry-collector-contrib/extension/httpf
 go 1.16
 
 require (
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
 	go.opentelemetry.io/collector/model v0.33.1-0.20210820002854-d3000232f8f6 // indirect
 	go.uber.org/zap v1.19.0
 )
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil => ../../internal/testutil
