+++ internal/aws/proxy/go.mod
@@ -7,4 +7,7 @@ require (
 	github.com/stretchr/testify v1.7.0
 	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
 	go.uber.org/zap v1.19.0
+    github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil v0.0.0-00010101000000-000000000000
 )
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil => ../internal/testutil
