+++ internal/testutil/go.mod
@@ -1,0 +1,5 @@
+module github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil
+
+go 1.16
+
+require github.com/stretchr/testify v1.7.0
