+++ receiver/carbonreceiver/go.mod
@@ -5,6 +5,7 @@ go 1.16
 require (
 	github.com/census-instrumentation/opencensus-proto v0.3.0
 	github.com/mattn/go-colorable v0.1.7 // indirect
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.7.0
 	go.opencensus.io v0.23.0
 	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
@@ -12,3 +13,5 @@ require (
 	go.uber.org/zap v1.19.0
 	google.golang.org/protobuf v1.27.1
 )
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil => ../../internal/testutil
