+++ receiver/sapmreceiver/go.mod
@@ -6,6 +6,7 @@ require (
 	github.com/gorilla/mux v1.8.0
 	github.com/jaegertracing/jaeger v1.25.0
 	github.com/open-telemetry/opentelemetry-collector-contrib/internal/splunk v0.0.0-00010101000000-000000000000
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil v0.0.0-00010101000000-000000000000
 	github.com/signalfx/sapm-proto v0.7.0
 	github.com/stretchr/testify v1.7.0
 	go.opencensus.io v0.23.0
@@ -13,5 +14,7 @@ require (
 	go.opentelemetry.io/collector/model v0.33.1-0.20210820002854-d3000232f8f6
 	go.uber.org/zap v1.19.0
 )
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil => ../../internal/testutil
 
 replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/splunk => ../../internal/splunk
