+++ receiver/statsdreceiver/go.mod
@@ -5,6 +5,7 @@ go 1.16
 require (
 	github.com/mattn/go-colorable v0.1.7 // indirect
 	github.com/montanaflynn/stats v0.0.0-20171201202039-1bf9dbcd8cbe
+	github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil v0.0.0-00010101000000-000000000000
 	github.com/stretchr/testify v1.7.0
 	go.opencensus.io v0.23.0
 	go.opentelemetry.io/collector v0.33.1-0.20210820002854-d3000232f8f6
@@ -12,3 +13,5 @@ require (
 	go.opentelemetry.io/otel v1.0.0-RC2
 	go.uber.org/zap v1.19.0
 )
+
+replace github.com/open-telemetry/opentelemetry-collector-contrib/internal/testutil => ../../internal/testutil
